Allow passing multiple contexts to ApiMarkdown::process()
Now it attempts to parse API links for each of the given contexts, and even if a context(s) is passed, it will try without a context as a last result.

Also: broken API links will no longer get any context added to them, since there's no way of knowing if the context was actually intended or not.